+++ spec/acceptance/juniper_spec.rb
@@ -2,11 +2,11 @@
 
 RSpec.describe 'Juniper' do
   subject(:client) do
-    WSDL::Client.new fixture('wsdl/juniper'), strictness: WSDL::Strictness.new(
+    WSDL::Client.new fixture('wsdl/juniper'), strictness: {
       schema_imports: false,
       schema_references: false,
       request_validation: false
-    )
+    }
   end
 
   it 'skips the relative schema import to still show other information' do